+++ assets/javascripts/kitten/components/molecules/buttons/button-image/index.js
@@ -47,7 +47,7 @@ const StyledButtonImage = styled.button`
   }
 
   &.k-ButtonImage--withBorder {
-    border: var(--border-width) solid var(--color-grey-400);
+    border: var(--border);
   }
 
   .k-ButtonImage__img {